chore: rename CrmUser button component
